requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.1
filetype==1.2.0
fonttools==4.61.1
fsspec==2025.12.0
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
scipy==1.15.3
six==1.17.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
ultralytics==8.3.245
ultralytics-thop==2.0.18
urllib3==2.6.2
